+++ Sources/SuperwallKit/Analytics/Superwall Event/SuperwallEventObjc.swift
@@ -116,6 +116,9 @@ public enum SuperwallEventObjc: Int, CaseIterable {
   /// When the loading of a paywall's website times out.
   case paywallWebviewLoadTimeout
 
+  /// When the loading of a paywall's website falls back to a different CDN.
+  case paywallWebviewLoadFallback
+
   /// When the request to load the paywall's products started.
   case paywallProductsLoadStart
 
@@ -124,6 +127,9 @@ public enum SuperwallEventObjc: Int, CaseIterable {
 
   /// When the request to load the paywall's products completed.
   case paywallProductsLoadComplete
+
+  /// When the request to load the paywall's products has failed and is being retried.
+  case paywallProductsLoadRetry
 
   /// Information about a paywall presentation request
   case paywallPresentationRequest
@@ -226,10 +232,14 @@ public enum SuperwallEventObjc: Int, CaseIterable {
       return "paywallWebviewLoad_complete"
     case .paywallWebviewLoadTimeout:
       return "paywallWebviewLoad_timeout"
+    case .paywallWebviewLoadFallback:
+      return "paywallWebviewLoad_fallback"
     case .paywallProductsLoadStart:
       return "paywallProductsLoad_start"
     case .paywallProductsLoadFail:
       return "paywallProductsLoad_fail"
+    case .paywallProductsLoadRetry:
+      return "paywallProductsLoad_retry"
     case .paywallProductsLoadComplete:
       return "paywallProductsLoad_complete"
     case .paywallPresentationRequest: