+++ Sources/SuperwallKit/Models/Paywall/Paywall.swift
@@ -4,7 +4,7 @@
 //
 //  Created by Yusuf Tör on 28/02/2022.
 //
-// swiftlint:disable function_body_length type_body_length
+// swiftlint:disable function_body_length type_body_length file_length
 
 import UIKit
 
@@ -22,8 +22,17 @@ struct Paywall: Decodable {
   /// The name of the paywall
   let name: String
 
+  /// The key used to cache the paywall object.
+  let cacheKey: String
+
+  /// The build ID of the Superwall configuration.
+  let buildId: String
+
   /// The URL of the paywall webpage
   var url: URL
+
+  /// An array of potential URLs to load the paywall from.
+  let urlConfig: WebViewURLConfig
 
   /// Contains the website modifications that are made on the paywall editor to be accepted
   /// by the webview.
@@ -113,7 +122,10 @@ struct Paywall: Decodable {
     case id
     case identifier
     case name
+    case cacheKey
+    case buildId
     case url
+    case urlConfig
     case htmlSubstitutions = "paywalljsEvent"
     case presentationStyle = "presentationStyleV2"
     case presentationCondition
@@ -146,7 +158,10 @@ struct Paywall: Decodable {
     databaseId = try values.decode(String.self, forKey: .id)
     identifier = try values.decode(String.self, forKey: .identifier)
     name = try values.decode(String.self, forKey: .name)
+    cacheKey = try values.decode(String.self, forKey: .cacheKey)
+    buildId = try values.decode(String.self, forKey: .buildId)
     url = try values.decode(URL.self, forKey: .url)
+    urlConfig = try values.decode(WebViewURLConfig.self, forKey: .urlConfig)
     htmlSubstitutions = try values.decode(String.self, forKey: .htmlSubstitutions)
 
     let throwableSurveys = try values.decodeIfPresent(
@@ -259,8 +274,11 @@ struct Paywall: Decodable {
     databaseId: String,
     identifier: String,
     name: String,
+    cacheKey: String,
+    buildId: String,
     experiment: Experiment? = nil,
     url: URL,
+    urlConfig: WebViewURLConfig,
     htmlSubstitutions: String,
     presentation: PaywallPresentationInfo,
     backgroundColorHex: String,
@@ -286,8 +304,11 @@ struct Paywall: Decodable {
     self.databaseId = databaseId
     self.identifier = identifier
     self.name = name
+    self.cacheKey = cacheKey
+    self.buildId = buildId
     self.experiment = experiment
     self.url = url
+    self.urlConfig = urlConfig
     self.htmlSubstitutions = htmlSubstitutions
     self.presentation = presentation
     self.backgroundColor = backgroundColor
@@ -317,6 +338,8 @@ struct Paywall: Decodable {
       databaseId: databaseId,
       identifier: identifier,
       name: name,
+      cacheKey: cacheKey,
+      buildId: buildId,
       url: url,
       products: products,
       productItems: productItems,
@@ -371,7 +394,10 @@ extension Paywall: Stubbable {
       databaseId: "id",
       identifier: "identifier",
       name: "abc",
+      cacheKey: "cacheKey",
+      buildId: "buildId",
       url: URL(string: "https://google.com")!,
+      urlConfig: .init(endpoints: [], maxAttempts: 0),
       htmlSubstitutions: "",
       presentation: PaywallPresentationInfo(
         style: .modal,