+++ Sources/SuperwallKit/Paywall/View Controller/PaywallViewController.swift
@@ -287,7 +287,7 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
           loadWebViewFromArchive(url: url)
         } else {
           // Fallback to old way if couldn't get archive
-          loadWebViewFromNetwork(url: url)
+          await webView.loadURL(from: paywall)
         }
       }
       return
@@ -296,25 +296,14 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
     if let webArchiveURL = paywallArchiveManager?.getCachedArchiveURL(manifest: paywall.manifest) {
       loadWebViewFromArchive(url: webArchiveURL)
     } else {
-      loadWebViewFromNetwork(url: url)
+      Task {
+        await webView.loadURL(from: paywall)
+      }
     }
   }
 
   private func loadWebViewFromArchive(url: URL) {
     webView.loadFileURL(url, allowingReadAccessTo: url)
-  }
-
-  private func loadWebViewFromNetwork(url: URL) {
-    if paywall.onDeviceCache == .enabled {
-      let request = URLRequest(
-        url: url,
-        cachePolicy: .returnCacheDataElseLoad
-      )
-      webView.load(request)
-    } else {
-      let request = URLRequest(url: url)
-      webView.load(request)
-    }
   }
 
   @objc private func reloadWebView() {
@@ -683,7 +672,7 @@ extension PaywallViewController {
       webView.setAllMediaPlaybackSuspended(false) // ignore-xcode-12
     }
 
-    if webView.didFailToLoad {
+    if webView.loadingHandler.didFailToLoad {
       loadWebView()
     }