+++ Tests/SuperwallKitTests/Paywall/View Controller/Web View/Message Handling/PaywallMessageHandlerTests.swift
@@ -18,8 +18,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -43,8 +43,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -69,8 +69,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -92,8 +92,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -117,8 +117,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -142,8 +142,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -167,8 +167,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -190,8 +190,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(
@@ -214,8 +214,8 @@ final class PaywallMessageHandlerTests: XCTestCase {
     )
     let webView = FakeWebView(
       isMac: false,
-      sessionEventsManager: dependencyContainer.sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: true,
       factory: dependencyContainer
     )
     let delegate = PaywallMessageHandlerDelegateMock(