+++ Tests/SuperwallKitTests/StoreKit/Products/ProductsFetcherSK1.swift
@@ -17,7 +17,11 @@ final class ProductsFetcherSK1Mock: ProductsFetcherSK1 {
     super.init()
   }
 
-  override func products(identifiers: Set<String>, forPaywall paywallName: String?) async throws -> Set<StoreProduct> {
+  override func products(
+    identifiers: Set<String>,
+    forPaywall paywall: Paywall?,
+    event: EventData?
+  ) async throws -> Set<StoreProduct> {
     return try productCompletionResult.get()
   }
 }